Splitting AddPublicationsModal into a Form + Modal
